pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "behaviorci"
version = "0.1.0"
description = "CI/CD for LLM behavior. Prompts don't ship until behavior passes tests."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "BehaviorCI Contributors" }
]
keywords = ["llm", "testing", "ci", "cd", "behavior", "prompts", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1,<9",
    "pydantic>=2.6,<3",
    "pyyaml>=6.0.1,<7",
    "httpx>=0.27,<1",
    "jsonschema>=4.21,<5",
    "rich>=13.7,<14",
    "jinja2>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5.0,<6",
    "ruff>=0.4,<1",
    "mypy>=1.10,<2",
    "types-PyYAML>=6.0,<7",
]

[project.scripts]
behaviorci = "behaviorci.cli:main"

[project.urls]
Homepage = "https://github.com/behaviorci/behaviorci"
Documentation = "https://github.com/behaviorci/behaviorci#readme"
Repository = "https://github.com/behaviorci/behaviorci"
Issues = "https://github.com/behaviorci/behaviorci/issues"

[tool.hatch.build.targets.wheel]
packages = ["behaviorci"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["behaviorci"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
